requirements.txt
flask
pandas
requests
gunicorn
python-dotenv
flask-cors
